Implement service in the controller
Define an implementation of the service and inject this service
implementation into the controller. In the controller, get items from
the service and pass it forward to the view.
In mustache template define an element that will represent an item.
Define elements with item's price and item's title. The test passes now.