Add Basket controller to add Items
The `form` in the view with send a POST request to the basket controller
. The basket controller will handle this request and will add an item to
the in-memory map buy the `sku` field as a key.
Create a Basket interface and an implementation of this interface. The
implementation is marked as a Spring component. But when we need to use
the basket service, then we inject the basket interface.