Mock Basket service w/o declaring class var
The Basket service is used in the CatalogController. But at the time of
the commit it is not used in the test. Spring Framework still requires
mock for all services used in the catalog. If the service is not needed
in the test, it can be mocked with class level annotation `@MockBeans`.
This way I don't need to define a class variable that won't be used.